Re-export with-source logging functions
The naming can be confusing, as one could assume that to add a source to
logInfo you might use logInfoS. And/or that logInfoNS means "No source".

We're attempting to respect upstream choices: logInfoNS means "Non-TH,
with Source". Logger.Aeson is mirroring Logger.CallStack and exposing
logInfo in place of logInfoN, so we are as well. This throws off the
natural symmetry between logInfoN/S in our interface.