Improve locating client ip in requestLogger
`Wai.remoteHost` is often localhost in proxied contexts. It's fine as a
final fallback, but we should be checking `X-Forwarded-For` or
`X-Real-IP` first. Adding another configuration point and improving the
default addresses this.

Fixes #18.

There aren't actual specifications for these headers, but they're
present in our AWS deployments and seem to behave as described here:

https://distinctplace.com/2014/04/23/story-behind-x-forwarded-for-and-x-real-ip-headers/

The client's IP address is appended to `X-Forwarded-For` each time a
proxy handles the request. Therefore, the very first value should be the
actual originating client's IP. `X-Real-Ip` _seems_ to be a late
addition to just provide the _final_ value from the forwarded-for again.

Trying these in turn should behave correctly in contexts that have these
headers.