Store LogSettings in the Logger
This can be useful in Apps that are going to make direct use of the
logger (e.g. `pushLoggerLn`) and need to know things like, "should I use
color?" By exposing the `LogSettings` directly, they can ensure
consistent behavior. If we didn't do this they may re-implement the
handling of `LogColorAuto` based on `LogDestination`, which is annoying
and error-prone. With this, they can simple ask any context with
`HasLogger`.