Add pushLogger and pushLoggerLn
We have a few apps that write both structured logging and unstructured
output. This is not usually a good idea, and there are flags to put that
output to a file if its to be processed, but having it in the logging is
useful when an operator just wants to see it. To ensure things are
ordered correctly, this output needs to go through the same `LoggerSet`
buffering. Allowing direct pushing is the only way to support that.