Add setLogSettingsConcurrency
Using fast-logger's default number of buffers can lead to reordered
logging output in the pursuit of performance. This is a good production
default, but CLIs can't tolerate it. This gives them a knob to tune down
the performance but ensure correct ordering (hopefully, I think).

Support for this depends on fast-logger version. 3.1.1 has all the
functions we need, 3.0.5 has it for logging to a file but not
`stdout`/`stderr`, and older than that has nothing. Because `LoggerSet`
construction is encapsulated, we can't implement this ourselves on the
older versions, so we just expose non-functional functions for that
case.