Add showLogLevels
I have a use-case where I'd like to pass the current App's `LOG_LEVEL`
onto a sub-process that also uses Blammo. Once we've read it into a
`LogLevels`, there was no way to faithfully reproduce it for the child
process without such a function.

We could re-organize things so we retain and pass it through as an
opaque `Text`, but that would be non-trivial in this case. Adding this
function seems reasonable enough and means we won't have to.