Don't log invalid details about a raw response
`wai` has a `RawResponse` constructor for `Response`[^1]. This is used
by applications that need to send raw bytes on the wire, such as
WebSockets.

This constructor has a field that is also type `Response`; functions
such as `responseStatus`[^2] or `responseHeaders` ultimately read from
this value since there is no actual status or headers that we can access
in the "raw" response itself.

This is an obvious but bad thing to do in something like `requestLogger`
because this inner response means nothing about what is actually
happening. It's there as a "fallback" to be used when raw responses
aren't supported. So it's almost always some kind of 500 error
response[^3]. Ultimately, this causes any Blammo-using applications to
report every WebSockets response as a 500 in the logs.

By its nature, there's not much we can do instead, as we know nothing
about the response, so this commit just changes things to omit response
details entirely for such cases. This is the approach taken by the more
official `Network.Wai.Middleware.RequestLogger` too.

[^1]: https://hackage.haskell.org/package/wai-3.2.4/docs/src/Network.Wai.Internal.html#Response
[^2]: https://hackage.haskell.org/package/wai-3.2.4/docs/src/Network.Wai.html#responseStatus
[^3]: https://hackage.haskell.org/package/yesod-core-1.6.26.0/docs/src/Yesod.Core.Handler.html#sendRawResponseNoConduit