Add setLoggerReformat
This refactors the contract between `reformatTerminal` and the
construction that happens in `newLogger`.

Previously, we constructed `lReformat` directly, which is a function of
only `LogLevel` and `ByteString`. This meant that a "reformat" function
such as `reformatTerminal` had to do some extra work that all such
functions would have to do:

- Construct the `Colors` value if needed
- Return that `ByteString` as-is if it doesn't parse as `LoggedMessage`

We can pull both of these responsibilities out into `newLogger` itself
and simplify `reformatTerminal` by giving it what it needs directly.

Also, by passing `LogSettings` entirely (instead of just the
breakpoint), we've now made its signature into something that works well
as a generic "reformat" interface.

Then, by putting that logic behind a `setLoggerReformat` that
`newLogger` uses internally, we give users a hook for assign any such
"reformat" function from the outside too:

```hs
main :: IO ()
main = do
  withLoggerEnv $ \logger -> do
    let app = App $ setLoggerReformat myReformat logger
    runAppM app -- ...

myReformat :: LogSettings -> Colors -> LogLevel -> LoggedMessage -> ByteString
myReformat = undefined
```

An extension like this doesn't interact with `LOG_FORMAT`, which may be
surprising, but it works well for situations that don't need to be
runtime-dynamic like that, such as CLIs.