Add support for adjusting colors through LogSettings
This commit extends our settings type with a `Colors -> Colors` function
(that can set through code, for now), which is called in all the
`getColors` functions.

This will allow end-user customization of `Colors` used in all logging.

Concretely, I want this so that I can fix for the fact that `dim`
doesn't work on GitHub Actions. This hook will allow me to easily
substitute `gray` which looks the same in that environment.