Swap gray for dim if running on GitHub Actions
GitHub Actions color support has a few limitations, not supporting `dim`
is one. Swapping in `gray` makes things looks as intended, but should
still be done conditionally because it won't look right elsewhere.

With the foundation of `setLogSettingsColors` and `endoOn` in place, it
was trivial to handle this for our users here, which is lower overall
complexity than making them do it themselves from the outside.